requirements.txt
Django==5.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-dateutil==2.8.2
channels==4.0.0
daphne==4.0.0
